ARM64/dts: mt6765: Disable unhandled or broken IRQ monitoring
This is primarily intended for generic x86 kernels which may run on all
sorts of broken systems. Our kernel will run on a very specific hardware
configuration with known drivers and devices, which makes this
unnecessary.

Enable noirqdebug to avoid the minor overhead added to every IRQ
handled.

eng.stk: adapt for sunfish

Suggested-by: Tyler Nijmeh <tylernij@gmail.com>
Change-Id: I8d55a7ffd884559bb4d3dd5d0d77c56b6645af73
Signed-off-by: kdrag0n <dragon@khronodragon.com>